requirements.txt
# Core dependencies
streamlit>=1.28.0
qdrant-client>=1.7.0
google-generativeai>=0.3.0
loguru>=0.7.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.0
tqdm>=4.66.0
pdfplumber>=0.9.0

# Optional
python-dotenv>=1.0.0 